Grill

+++ code/modules/cooking_with_jane/cooking_tracker.dm
@@ -124,14 +124,17 @@
 
 	if(valid_steps.len > 1)
 		completion_lockout = TRUE
-		var/list/choice = input(user, "There's two things you can do with this item!", "Choose One:") in valid_steps
-		completion_lockout = FALSE
-		if(!choice)
-			#ifdef CWJ_DEBUG
-			log_debug("/recipe_tracker/proc/process_item returned choice cancel!")
-			#endif
-			return CWJ_CHOICE_CANCEL
-		use_class = choice
+		if(user)
+			var/list/choice = input(user, "There's two things you can do with this item!", "Choose One:") in valid_steps
+			completion_lockout = FALSE
+			if(!choice)
+				#ifdef CWJ_DEBUG
+				log_debug("/recipe_tracker/proc/process_item returned choice cancel!")
+				#endif
+				return CWJ_CHOICE_CANCEL
+			use_class = choice
+		else
+			use_class = 1
 	#ifdef CWJ_DEBUG
 	log_debug("Use class determined: [use_class]")
 	#endif
@@ -151,7 +154,7 @@
 				used_id = TRUE
 				if(step.is_complete(used_object, src))
 					has_traversed = TRUE
-					pointer.traverse(step.unique_id)
+					pointer.traverse(step.unique_id, used_object)
 					break
 		if (!used_id)
 			active_recipe_pointers.Remove(pointer)
@@ -169,13 +172,13 @@
 				recipe_string = "\a [pointer.current_recipe.name]"
 			else
 				recipe_string += ", or \a [pointer.current_recipe.name]"
-
-		if(alert(user, "If you finish cooking now, you will create [recipe_string]. However, you feel there are possibilities beyond even this. Continue cooking anyways?",,"Yes","No") == "Yes")
-			//Cull finished recipe items
-			for (var/datum/cooking_with_jane/recipe_pointer/pointer in completed_list)
-				active_recipe_pointers.Remove(pointer)
-				qdel(pointer)
-			completed_list = list()
+		if(user)
+			if(alert(user, "If you finish cooking now, you will create [recipe_string]. However, you feel there are possibilities beyond even this. Continue cooking anyways?",,"Yes","No") == "Yes")
+				//Cull finished recipe items
+				for (var/datum/cooking_with_jane/recipe_pointer/pointer in completed_list)
+					active_recipe_pointers.Remove(pointer)
+					qdel(pointer)
+				completed_list = list()
 
 	//Check if we completed our recipe
 	var/datum/cooking_with_jane/recipe_pointer/chosen_pointer = null
@@ -184,12 +187,13 @@
 		log_debug("/recipe_tracker/proc/process_item YO WE ACTUALLY HAVE A COMPLETED A RECIPE!")
 		#endif
 		chosen_pointer = completed_list[1]
-		if(completed_list.len > 1)
-			completion_lockout = TRUE
-			var/choice = input(user, "There's two things you complete at this juncture!", "Choose One:") in completed_list
-			completion_lockout = FALSE
-			if(choice)
-				chosen_pointer = completed_list[choice]
+		if(user)
+			if(completed_list.len > 1)
+				completion_lockout = TRUE
+				var/choice = input(user, "There's two things you complete at this juncture!", "Choose One:") in completed_list
+				completion_lockout = FALSE
+				if(choice)
+					chosen_pointer = completed_list[choice]
 
 	//Call a proc that follows one of the steps in question, so we have all the nice to_chat calls.
 	var/datum/cooking_with_jane/recipe_step/sample_step = valid_steps[1]
